+++ CHANGELOG.md
@@ -3,10 +3,14 @@ All notable changes to this project will be documented in this file based on the
 [Keep a Changelog](http://keepachangelog.com/) Standard.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased](https://github.com/rokka-io/rokka-client-php-cli/compare/1.6.1...master)
+## [Unreleased](https://github.com/rokka-io/rokka-client-php-cli/compare/1.6.2...master)
 
 ### Added
+-  `organization:membership:add` takes now multiple roles, instead of just one
+
 ### Changed
+- Adjusted to new rokka/client 1.7.0 membership methods
+- Membership methods take a user_id instead of an email now (email still works, but is deprecated)
 ### Deprecated
 ### Fixed
 ### Removed